backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=2.4.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
python-dotenv>=0.19.0
# LangChain
langchain
langchain_community
langchain_core
chromadb>=0.5.0
langchain-text-splitters
langchain-huggingface
requests>=2.31.0
langchain-chroma
pydantic>=1.8.0,<2.0.0
pymongo>=3.12.0
sentence-transformers